+++ resources/views/components/layout.blade.php
@@ -2,7 +2,8 @@
 <html lang="en" class="h-full bg-gray-100">
 <head>
     <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
+    <meta name="viewport"
+          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>Home page</title>
     @vite('resources/css/app.css')
@@ -17,7 +18,9 @@
                 <div class="flex">
                     <div class="shrink-0 flex items-center">
                         <a href="{{ route('home') }}">
-                            <img class="h-8" src="https://api.logo.com/api/v2/images?logo=lg_RdPmxPeu7W4uYi1t17&format=webp&width=2000&background=transparent&fit=contain&quality=100&u=2024-10-16T20%3A20%3A18.823Z" alt="Pawfect Match">
+                            <img class="h-8"
+                                 src="https://api.logo.com/api/v2/images?logo=lg_RdPmxPeu7W4uYi1t17&format=webp&width=2000&background=transparent&fit=contain&quality=100&u=2024-10-16T20%3A20%3A18.823Z"
+                                 alt="Pawfect Match">
                         </a>
                     </div>
 
@@ -41,12 +44,19 @@
                                 <x-dropdown-link :href="route('profile.edit')">
                                     {{ __('Profile') }}
                                 </x-dropdown-link>
-                                <!-- authenticatie -->
+
+                                {{--admin dashboard link--}}
+                                @if(Auth::user()->role === 'admin')
+                                    <x-dropdown-link :href="route('dashboard')">
+                                        {{ __('Dashboard') }}
+                                    </x-dropdown-link>
+                                @endif
+
+                                <!-- logout -->
                                 <form method="POST" action="{{ route('logout') }}">
                                     @csrf
                                     <x-dropdown-link :href="route('logout')"
-                                                     onclick="event.preventDefault();
-                                                this.closest('form').submit();">
+                                                     onclick="event.preventDefault(); this.closest('form').submit();">
                                         {{ __('Log Out') }}
                                     </x-dropdown-link>
                                 </form>
@@ -66,7 +76,8 @@
                             <path :class="{'hidden': open, 'inline-flex': ! open }" class="inline-flex"
                                   stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
                                   d="M4 6h16M4 12h16M4 18h16"/>
-                            <path :class="{'hidden': ! open, 'inline-flex': open }" class="hidden" stroke-linecap="round"
+                            <path :class="{'hidden': ! open, 'inline-flex': open }" class="hidden"
+                                  stroke-linecap="round"
                                   stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12"/>
                         </svg>
                     </button>
@@ -86,10 +97,12 @@
             <div class="pt-4 pb-1 border-t border-gray-200 dark:border-gray-600">
                 <div class="px-4">
                     @if(Auth::check())
-                        <div class="font-medium text-base text-gray-800 dark:text-gray-200">{{ Auth::user()->name }}</div>
+                        <div
+                            class="font-medium text-base text-gray-800 dark:text-gray-200">{{ Auth::user()->name }}</div>
                         <div class="font-medium text-sm text-gray-500">{{ Auth::user()->email }}</div>
                     @else
-                        <a href="{{ route('login') }}" class="text-sm text-gray-700 dark:text-gray-500 underline">Login</a>
+                        <a href="{{ route('login') }}"
+                           class="text-sm text-gray-700 dark:text-gray-500 underline">Login</a>
                     @endif
                 </div>
                 <div class="mt-3 space-y-1">
@@ -108,18 +121,19 @@
                             </x-responsive-nav-link>
                         </form>
                     @else
-                        <x-responsive-nav-link href="/register" :active="request()->is('register')">Register</x-responsive-nav-link>
+                        <x-responsive-nav-link href="/register" :active="request()->is('register')">Register
+                        </x-responsive-nav-link>
                     @endif
                 </div>
             </div>
         </div>
     </nav>
 
-{{--    <header class="bg-white shadow">--}}
-{{--        <div class="mx-auto max-w-7xl px-4 py-6 sm:px-6 lg:px-8">--}}
-{{--            <h1 class="text-3xl font-bold tracking-tight text-gray-900"></h1>--}}
-{{--        </div>--}}
-{{--    </header>--}}
+    {{--    <header class="bg-white shadow">--}}
+    {{--        <div class="mx-auto max-w-7xl px-4 py-6 sm:px-6 lg:px-8">--}}
+    {{--            <h1 class="text-3xl font-bold tracking-tight text-gray-900"></h1>--}}
+    {{--        </div>--}}
+    {{--    </header>--}}
 
     <main>
         <div class="mx-auto max-w-7xl px-4 py-6 sm:px-6 lg:px-8">